util for external pv
prune unused pv utils and its test
